Added support for MPI runs on frontier

+++ build-papi.sh
@@ -11,23 +11,30 @@ export PAPI_ROOT=$INSTALL_DIR/papi
 export PAPI_LIB=$PAPI_ROOT/lib
 
 cd $BUILD_DIR
-rm -Rf rocm_smi_lib
-git clone https://github.com/adam-mcdaniel/rocm_smi_lib
-
-cd $BUILD_DIR/rocm_smi_lib/
-rm -Rf build
-mkdir build
-cd build
-cmake .. -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$INSTALL_DIR/rocm_smi_lib
-make -j16 install
+if [ ! -d "rocm_smi_lib" ]; then
+    echo "Cloning rocm_smi_lib repository..."
+    rm -Rf rocm_smi_lib
+    git clone https://github.com/adam-mcdaniel/rocm_smi_lib
+else
+    cd $BUILD_DIR/rocm_smi_lib/
+    rm -Rf build
+    mkdir build
+    cd build
+    cmake .. -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$INSTALL_DIR/rocm_smi_lib
+    make -j16 install
+fi
 
 cd $BUILD_DIR
-rm -Rf papi
-git clone https://github.com/adam-mcdaniel/papi
-
-# Install PAPI
-cd $BUILD_DIR/papi/src
-make clean
-./configure --with-components="coretemp rocm_smi" --prefix $INSTALL_DIR/papi
-make -j16
-make install-all
+if [ ! -d "papi" ]; then
+    echo "Cloning PAPI repository..."
+    rm -Rf papi
+    git clone https://github.com/adam-mcdaniel/papi
+else
+    echo "PAPI repository already exists, skipping clone."
+    # Install PAPI
+    cd $BUILD_DIR/papi/src
+    make clean
+    ./configure --with-components="coretemp rocm_smi" --prefix $INSTALL_DIR/papi
+    make -j16
+    make install-all
+fi
